PARROTS!!
This is a very foundational start for having parrots added in game.  This is affected by similar movment problems that all Entities have and not all features work as of yet.

Known issues:

Parrots do not 'dance' when near jukeboxes.
Parrots do not die when you attempt to feed them cookies.
Parrots 'walk' while flying.
Parrots do not spawn naturally.