Code cleanup (#251)
* Change to Log PreText

Updates log pretext to make it obvious this is from PureEntitiesX

* Implement Panic Trait

This creates default methods and variables for Mobs that use IntfCanPanic.

* Lots of Cleanup

This update includes many updates to cleanup duplicated code and implement traits that reduce code duplication.  Also several grammer updates and a few other minor items.

Fixes drops on Zombie variants to be closer to vanilla.

Begin implemnting Slimes and Magma Cubes

Update CustomLogger namespace to be unique to PureEntitiesX

* Remove Individual Constants for Newtork IDs

This removes the individual constants in Data.php that were previously used for Network IDs.  This has been replaced with the NetworkIDs array.

* Revert "Remove Individual Constants for Newtork IDs"

This reverts commit b38f674c250198733653a58d433ae662aa0026c1.

* Reformat to Match provided formatting file.

* More Reformatting.

* Revert File Structure Changes

This should revert changes made in PR #237 to the directory structure.

* Fix Poggit for new directory structure.